Refactor UnsafeConnectionBuilder functionality
